Wrap logging functions, move GetLoggerFromStackTrace into ResoniteMod class
